requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
pydantic>=2.0.0
motor>=3.3.0
python-dotenv>=1.0.0
dnspython>=2.0.0
